+++ semTools/R/probeInteraction.R
@@ -1,7 +1,11 @@
 ### Sunthud Pornprasertmanit & Terrence D. Jorgensen
-### Last updated: 9 April 2020
+### Last updated: 9 January 2021
 
 
+
+## --------
+## 2-way MC
+## --------
 
 ##' Probing two-way interaction on the no-centered or mean-centered latent
 ##' interaction
@@ -318,9 +322,9 @@ probe2WayMC <- function(fit, nameX, nameY, modVar, valProbe, group,
 		usedBeta <- betaNC
 
 		# Change the order of usedVar and usedBeta if the moderator variable is listed first
-		if(modVar == 2) {
+		if(modVar == 1) {
 			usedVar <- usedVar[c(2, 1, 3), c(2, 1, 3)]
-			# usedBeta <- usedBeta[c(2, 1, 3)]
+			usedBeta <- usedBeta[c(2, 1, 3)]
 		}
 
 		# Find simple slope
@@ -336,6 +340,10 @@ probe2WayMC <- function(fit, nameX, nameY, modVar, valProbe, group,
 }
 
 
+
+## --------
+## 2-way RC
+## --------
 
 ##' Probing two-way interaction on the residual-centered latent interaction
 ##'
@@ -448,7 +456,7 @@ probe2WayMC <- function(fit, nameX, nameY, modVar, valProbe, group,
 ##'
 ##' Geldhof, G. J., Pornprasertmanit, S., Schoemann, A. M., & Little, T. D.
 ##' (2013). Orthogonalizing through residual centering: Extended applications
-##' and caveats \emph{Educational and Psychological Measurement, 73}(1), 27--46.
+##' and caveats. \emph{Educational and Psychological Measurement, 73}(1), 27--46.
 ##' doi:10.1177/0013164412445473
 ##' @examples
 ##'
@@ -690,9 +698,9 @@ probe2WayRC <- function(fit, nameX, nameY, modVar, valProbe, group,
 		usedBeta <- betaNC
 
 		# Change the order of usedVar and usedBeta if the moderator variable is listed first
-		if(modVar == 2) {
+		if (modVar == 1) {
 			usedVar <- usedVar[c(2, 1, 3), c(2, 1, 3)]
-			# usedBeta <- usedBeta[c(2, 1, 3)]
+			usedBeta <- usedBeta[c(2, 1, 3)]
 		}
 
 		# Find simple slope
@@ -708,6 +716,10 @@ probe2WayRC <- function(fit, nameX, nameY, modVar, valProbe, group,
 }
 
 
+
+## --------
+## 3-way MC
+## --------
 
 ##' Probing three-way interaction on the no-centered or mean-centered latent
 ##' interaction
@@ -1088,6 +1100,10 @@ probe3WayMC <- function(fit, nameX, nameY, modVar, valProbe1, valProbe2, group,
 }
 
 
+
+## --------
+## 3-way RC
+## --------
 
 ##' Probing three-way interaction on the residual-centered latent interaction
 ##'
@@ -1545,6 +1561,10 @@ probe3WayRC <- function(fit, nameX, nameY, modVar, valProbe1, valProbe2, group,
 }
 
 
+
+## -----------------
+## Plotting Function
+## -----------------
 
 ##' Plot a latent interaction
 ##'