+++ semTools/NEWS.md
@@ -4,13 +4,13 @@
 ## New Features:
 
 - `summary()` method for `lavaan.mi` objects now passes arguments to `lavTestLRT.mi()` via `...`.
-- `emmeans` methods deal with models fit with missing data.
 
 ## Bug Fixes:
 
 - `reliability()` incorrectly used standardized thresholds rather than the model's parameter estimates (i.e., on the latent-response scales) to calculate omega for categorical indicators.
 - `net()` returned an error for models with categorical indicators, now fixed.
 - The wrong *df* were used to find confidence limits for RMSEA calculated from the scaled $\chi^2$ statistic in `lavaan.mi` objects.
+- `emmeans` methods appropriately deal with incomplete data.
 
 
 # semTools 0.5-3 (on CRAN 27 May 2020)